requirements.txt
av
pillow
numpy
tqdm
argparse
torch==0.4.1
torchvision==0.2.2